health_assistant/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sentence-transformers==3.0.0
faiss-cpu==1.12.0
transformers==4.44.0
googletrans==4.0.0-rc1
langdetect==1.0.9
numpy==1.26.4
torch==2.9.0